Move mappings to separate entry points
