+++ angular-workspace/projects/ni/nimble-angular/mapping/icon/nimble-mapping-icon.directive.ts
