+++ angular-workspace/projects/ni/nimble-angular/mapping/icon/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-mapping-icon.directive';
+export * from './nimble-mapping-icon.module';
