+++ angular-workspace/projects/ni/nimble-angular/mapping/spinner/nimble-mapping-spinner.directive.ts
