+++ angular-workspace/projects/ni/nimble-angular/mapping/spinner/nimble-mapping-spinner.module.ts
