+++ angular-workspace/projects/ni/nimble-angular/mapping/spinner/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-mapping-spinner.directive';
+export * from './nimble-mapping-spinner.module';
